# stock-serveur-v2/target/swagger.json
{
  "swagger" : "2.0",
  "info" : {
    "description" : "Swagger API Documentation for gestion stock application",
    "version" : "0.0.1-SNAPSHOT",
    "title" : "Swagger API Documentation for Gestion de stock application",
    "contact" : {
      "name" : "Samuel Vermon Ngnawen",
      "email" : "ngnawens@gmail.com"
    }
  },
  "basePath" : "http://localhost:8081",
...
  "paths" : {
...
    "/api2/articles/{id}" : {
      "get" : {
        "tags" : [ "api2articles" ],
        "summary" : "Recherche d'un article",
        "description" : "Cette methode permet de rechercher un article par son ID",
        "operationId" : "findById",
        "produces" : [ "application/json" ],
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "type" : "integer",
          "format" : "int64"
        } ],
        "responses" : {
          "200" : {
            "description" : "Article a été retrouvé dns la BD",
            "schema" : {
              "$ref" : "#/definitions/ArticleDto"
            }
          },
          "404" : {
            "description" : "Aucun article n'existe dans la BD avec l'ID fourni"
          }
        }
      },
      "delete" : {
        "tags" : [ "api2articles" ],
        "summary" : "Suppression d'un article",
        "description" : "Cette methode permet de supprimer un article à partir de son ID",
        "operationId" : "delete",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "type" : "integer",
          "format" : "int64"
        } ],
        "responses" : {
          "200" : {
            "description" : "Article supprimé avec succès",
            "schema" : {
              "$ref" : "#/definitions/ArticleDto"
            }
          }
        }
      }
    },
    "/api2/authenticate/authentication" : {
      "post" : {
        "tags" : [ "api2authenticate" ],
        "operationId" : "authentication",
        "parameters" : [ {
          "in" : "body",
          "name" : "body",
          "required" : false,
          "schema" : {
            "$ref" : "#/definitions/AuthenticationRequest"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "$ref" : "#/definitions/AuthenticationResponse"
            }
          }
        }
      }
    },
...
    "/api2/categories/{id}" : {
      "get" : {
        "tags" : [ "api2categories" ],
        "summary" : "Recherche d'une categorie",
        "description" : "Cette methode permet de rechercher une categorie par son ID",
        "operationId" : "findById",
        "produces" : [ "application/json" ],
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "type" : "integer",
          "format" : "int64"
        } ],
        "responses" : {
          "200" : {
            "description" : "Categorie a été retrouvée dans la BD",
            "schema" : {
              "$ref" : "#/definitions/CategorieDto"
            }
          },
          "404" : {
            "description" : "Aucune categorie n'existe dans la BD avec l'ID fourni"
          }
        }
      },
      "delete" : {
        "tags" : [ "api2categories" ],
        "summary" : "Suppression d'une categorie",
        "description" : "Cette methode permet de supprimer une categories à partir de son ID",
        "operationId" : "delete",
        "responses" : {
          "200" : {
            "description" : "Categorie supprimé avec succès",
            "schema" : {
              "$ref" : "#/definitions/CategorieDto"
            }
          }
        }
      }
    },
...
    "/api2/commandeClients/{id}" : {
      "get" : {
        "tags" : [ "api2commandesclients" ],
        "operationId" : "findById",
        "produces" : [ "application/json" ],
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "type" : "integer",
          "format" : "int64"
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "$ref" : "#/definitions/CommandeClientDto"
            }
          }
        }
      },
      "delete" : {
        "tags" : [ "api2commandesclients" ],
        "operationId" : "delete",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "type" : "integer",
          "format" : "int64"
        } ],
        "responses" : {
          "default" : {
            "description" : "successful operation"
          }
        }
      }
    },
...
    "/api2/commandeFournisseurs/{id}" : {
      "get" : {
        "tags" : [ "api2commandesfournisseurs" ],
        "operationId" : "findById",
        "produces" : [ "application/json" ],
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "type" : "integer",
          "format" : "int64"
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "$ref" : "#/definitions/CommandeFournisseurDto"
            }
          }
        }
      },
      "delete" : {
        "operationId" : "delete",
        "responses" : {
          "default" : {
            "description" : "successful operation"
          }
        }
      }
    },
...
    "/api2/roles/{id}" : {
      "get" : {
        "tags" : [ "api2roles" ],
        "operationId" : "findById",
        "produces" : [ "application/json" ],
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "type" : "integer",
          "format" : "int64"
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "$ref" : "#/definitions/RolesDto"
            }
          }
        }
      },
      "delete" : {
        "tags" : [ "api2roles" ],
        "operationId" : "delete",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "type" : "integer",
          "format" : "int64"
        } ],
        "responses" : {
          "default" : {
            "description" : "successful operation"
          }
        }
      }
    },
    "/api2/sousCategories" : {
      "get" : {
        "tags" : [ "api2souscategories" ],
        "operationId" : "findAll",
        "produces" : [ "application/json" ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "type" : "array",
              "items" : {
                "$ref" : "#/definitions/SousCategorieDto"
              }
            }
          }
        }
      },
      "post" : {
        "tags" : [ "api2souscategories" ],
        "operationId" : "save",
        "consumes" : [ "application/json" ],
        "produces" : [ "application/json" ],
        "parameters" : [ {
          "in" : "body",
          "name" : "body",
          "required" : false,
          "schema" : {
            "$ref" : "#/definitions/SousCategorieDto"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "$ref" : "#/definitions/SousCategorieDto"
            }
          }
        }
      }
    },
...
    "/api2/sousCategories/{id}" : {
      "get" : {
        "tags" : [ "api2souscategories" ],
        "operationId" : "findById",
        "produces" : [ "application/json" ],
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "type" : "integer",
          "format" : "int64"
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "$ref" : "#/definitions/SousCategorieDto"
            }
          }
        }
      },
      "delete" : {
        "tags" : [ "api2souscategories" ],
        "operationId" : "delete",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "type" : "integer",
          "format" : "int64"
        } ],
        "responses" : {
          "default" : {
            "description" : "successful operation"
          }
        }
      }
    },
...
    "/api2/users/{id}" : {
      "get" : {
        "tags" : [ "api2users" ],
        "operationId" : "findById",
        "produces" : [ "application/json" ],
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "type" : "integer",
          "format" : "int64"
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "$ref" : "#/definitions/UsersDto"
            }
          }
        }
      },
      "delete" : {
        "tags" : [ "api2users" ],
        "operationId" : "delete",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "type" : "integer",
          "format" : "int64"
        } ],
        "responses" : {
          "default" : {
            "description" : "successful operation"
          }
        }
      }
    },
    "/api2/ventes" : {
      "get" : {
        "tags" : [ "api2ventes" ],
        "operationId" : "findAll",
        "produces" : [ "application/json" ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "type" : "array",
              "items" : {
                "$ref" : "#/definitions/VenteDto"
              }
            }
          }
        }
      },
      "post" : {
        "tags" : [ "api2ventes" ],
        "operationId" : "save",
        "consumes" : [ "application/json" ],
        "produces" : [ "application/json" ],
        "parameters" : [ {
          "in" : "body",
          "name" : "body",
          "required" : false,
          "schema" : {
            "$ref" : "#/definitions/VenteDto"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "$ref" : "#/definitions/VenteDto"
            }
          }
        }
      }
    },
    "/api2/ventes/{code}" : {
      "get" : {
        "tags" : [ "api2ventes" ],
        "operationId" : "findVenteByCode",
        "produces" : [ "application/json" ],
        "parameters" : [ {
          "name" : "code",
          "in" : "path",
          "required" : true,
          "type" : "string"
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "$ref" : "#/definitions/VenteDto"
            }
          }
        }
      }
    },
    "/api2/ventes/{id}" : {
      "get" : {
        "operationId" : "findById",
        "produces" : [ "application/json" ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "$ref" : "#/definitions/VenteDto"
            }
          }
        }
      },
      "delete" : {
        "tags" : [ "api2ventes" ],
        "operationId" : "delete",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "type" : "integer",
          "format" : "int64"
        } ],
        "responses" : {
          "default" : {
            "description" : "successful operation"
          }
        }
      }
    },
...
  },
  "definitions" : {
    "Adresse" : {
      "type" : "object",
      "properties" : {
        "adresse1" : {
          "type" : "string"
        },
        "adresse2" : {
          "type" : "string"
        },
        "ville" : {
          "type" : "string"
        },
        "codePostale" : {
          "type" : "string"
        },
        "pays" : {
          "type" : "string"
        }
      }
    },
    "AdresseDto" : {
      "type" : "object",
      "properties" : {
        "adresse1" : {
          "type" : "string"
        },
        "adresse2" : {
          "type" : "string"
        },
        "ville" : {
          "type" : "string"
        },
        "codePostale" : {
          "type" : "string"
        },
        "pays" : {
          "type" : "string"
        }
      }
    },
    "Article" : {
      "type" : "object",
      "properties" : {
        "createdBy" : {
          "type" : "string"
        },
        "updatedBy" : {
          "type" : "string"
        },
        "createdDate" : {
          "type" : "string",
          "format" : "date-time"
        },
        "updateDate" : {
          "type" : "string",
          "format" : "date-time"
        },
        "id" : {
          "type" : "integer",
          "format" : "int64"
        },
        "code" : {
          "type" : "string"
        },
        "designation" : {
          "type" : "string"
        },
        "prixUnitaireHt" : {
          "type" : "number"
        },
        "tauxTva" : {
          "type" : "number"
        },
        "prixUnitaireTtc" : {
          "type" : "number"
        },
        "prixAchat" : {
          "type" : "number"
        },
        "prixVente" : {
          "type" : "number"
        },
        "quantite" : {
          "type" : "number",
          "format" : "double"
        },
        "photo" : {
          "type" : "string"
        },
        "categorie" : {
          "$ref" : "#/definitions/Categorie"
        },
        "entrepriseId" : {
          "type" : "integer",
          "format" : "int64"
        }
      }
    },
    "ArticleDto" : {
      "type" : "object",
      "properties" : {
        "id" : {
          "type" : "integer",
          "format" : "int64"
        },
        "code" : {
          "type" : "string"
        },
        "designation" : {
          "type" : "string"
        },
        "prixUnitaireHt" : {
          "type" : "number"
        },
        "tauxTva" : {
          "type" : "number"
        },
        "prixUnitaireTtc" : {
          "type" : "number"
        },
        "prixAchat" : {
          "type" : "number"
        },
        "prixVente" : {
          "type" : "number"
        },
        "quantite" : {
          "type" : "number",
          "format" : "double"
        },
        "photo" : {
          "type" : "string"
        },
        "categorieDto" : {
          "$ref" : "#/definitions/CategorieDto"
        },
        "entrepriseId" : {
          "type" : "integer",
          "format" : "int64"
        }
      }
    },
    "AuthenticationRequest" : {
      "type" : "object",
      "properties" : {
        "login" : {
          "type" : "string"
        },
        "password" : {
          "type" : "string"
        }
      }
    },
    "AuthenticationResponse" : {
      "type" : "object",
      "properties" : {
        "accessToken" : {
          "type" : "string"
        }
      }
    },
    "Categorie" : {
      "type" : "object",
      "properties" : {
        "createdBy" : {
          "type" : "string"
        },
        "updatedBy" : {
          "type" : "string"
        },
        "createdDate" : {
          "type" : "string",
          "format" : "date-time"
        },
        "updateDate" : {
          "type" : "string",
          "format" : "date-time"
        },
        "id" : {
          "type" : "integer",
          "format" : "int64"
        },
        "code" : {
          "type" : "string"
        },
        "libelle" : {
          "type" : "string"
        },
        "entrepriseId" : {
          "type" : "integer",
          "format" : "int64"
        },
        "articleList" : {
          "type" : "array",
          "items" : {
            "$ref" : "#/definitions/Article"
          }
        }
      }
    },
    "CategorieDto" : {
      "type" : "object",
      "properties" : {
        "id" : {
          "type" : "integer",
          "format" : "int64"
        },
        "code" : {
          "type" : "string"
        },
        "libelle" : {
          "type" : "string"
        },
        "entrepriseId" : {
          "type" : "integer",
          "format" : "int64"
        }
      }
    },
    "Client" : {
      "type" : "object",
      "properties" : {
        "createdBy" : {
          "type" : "string"
        },
        "updatedBy" : {
          "type" : "string"
        },
        "createdDate" : {
          "type" : "string",
          "format" : "date-time"
        },
        "updateDate" : {
          "type" : "string",
          "format" : "date-time"
        },
        "id" : {
          "type" : "integer",
          "format" : "int64"
        },
        "nom" : {
          "type" : "string"
        },
        "prenom" : {
          "type" : "string"
        },
        "genre" : {
          "type" : "string"
        },
        "dateNaissance" : {
          "type" : "string",
          "format" : "date"
        },
        "lieuNaissance" : {
          "type" : "string"
        },
        "email" : {
          "type" : "string"
        },
        "numTel" : {
          "type" : "string"
        },
        "photo" : {
          "type" : "string"
        },
        "adresse" : {
          "$ref" : "#/definitions/Adresse"
        },
        "commandeClientList" : {
          "type" : "array",
          "items" : {
            "$ref" : "#/definitions/CommandeClient"
          }
        },
        "entrepriseId" : {
          "type" : "integer",
          "format" : "int64"
        }
      }
    },
...
    "CommandeClient" : {
      "type" : "object",
      "properties" : {
        "createdBy" : {
          "type" : "string"
        },
        "updatedBy" : {
          "type" : "string"
        },
        "createdDate" : {
          "type" : "string",
          "format" : "date-time"
        },
        "updateDate" : {
          "type" : "string",
          "format" : "date-time"
        },
        "id" : {
          "type" : "integer",
          "format" : "int64"
        },
        "code" : {
          "type" : "string"
        },
        "dateCommande" : {
          "type" : "integer",
          "format" : "int64"
        },
        "client" : {
          "$ref" : "#/definitions/Client"
        },
        "ligneCommandeClientList" : {
          "type" : "array",
          "items" : {
            "$ref" : "#/definitions/LigneCommandeClient"
          }
        },
        "entrepriseId" : {
          "type" : "integer",
          "format" : "int64"
        }
      }
    },
    "CommandeClientDto" : {
      "type" : "object",
      "properties" : {
        "id" : {
          "type" : "integer",
          "format" : "int64"
        },
        "code" : {
          "type" : "string"
        },
        "dateCommande" : {
          "type" : "integer",
          "format" : "int64"
        },
        "client" : {
          "$ref" : "#/definitions/Client"
        },
        "entrepriseId" : {
          "type" : "integer",
          "format" : "int64"
        },
        "ligneCommandeClientList" : {
          "type" : "array",
          "items" : {
            "$ref" : "#/definitions/LigneCommandeClientDto"
          }
        }
      }
    },
    "CommandeFournisseur" : {
      "type" : "object",
      "properties" : {
        "createdBy" : {
          "type" : "string"
        },
        "updatedBy" : {
          "type" : "string"
        },
        "createdDate" : {
          "type" : "string",
          "format" : "date-time"
        },
        "updateDate" : {
          "type" : "string",
          "format" : "date-time"
        },
        "id" : {
          "type" : "integer",
          "format" : "int64"
        },
        "code" : {
          "type" : "string"
        },
        "dateCommande" : {
          "type" : "integer",
          "format" : "int64"
        },
        "fournisseur" : {
          "$ref" : "#/definitions/Fournisseur"
        },
        "ligneCommandeFournisseurList" : {
          "type" : "array",
          "items" : {
            "$ref" : "#/definitions/LigneCommandeFournisseur"
          }
        },
        "entrepriseId" : {
          "type" : "integer",
          "format" : "int64"
        }
      }
    },
    "CommandeFournisseurDto" : {
      "type" : "object",
      "properties" : {
        "id" : {
          "type" : "integer",
          "format" : "int64"
        },
        "code" : {
          "type" : "string"
        },
        "dateCommande" : {
          "type" : "integer",
          "format" : "int64"
        },
        "fournisseur" : {
          "$ref" : "#/definitions/Fournisseur"
        },
        "entrepriseId" : {
          "type" : "integer",
          "format" : "int64"
        },
        "ligneCommandeFournisseurList" : {
          "type" : "array",
          "items" : {
            "$ref" : "#/definitions/LigneCommandeFournisseurDto"
          }
        }
      }
    },
    "EntrepriseDto" : {
      "type" : "object",
      "properties" : {
        "id" : {
          "type" : "integer",
          "format" : "int64"
        },
        "nom" : {
          "type" : "string"
        },
        "logo" : {
          "type" : "string"
        }
      }
    },
    "Fournisseur" : {
      "type" : "object",
      "properties" : {
        "createdBy" : {
          "type" : "string"
        },
        "updatedBy" : {
          "type" : "string"
        },
        "createdDate" : {
          "type" : "string",
          "format" : "date-time"
        },
        "updateDate" : {
          "type" : "string",
          "format" : "date-time"
        },
        "id" : {
          "type" : "integer",
          "format" : "int64"
        },
        "nom" : {
          "type" : "string"
        },
        "prenom" : {
          "type" : "string"
        },
        "genre" : {
          "type" : "string"
        },
        "dateNaissance" : {
          "type" : "integer",
          "format" : "int64"
        },
        "lieuNaissance" : {
          "type" : "string"
        },
        "email" : {
          "type" : "string"
        },
        "numTel" : {
          "type" : "string"
        },
        "photo" : {
          "type" : "string"
        },
        "adresse" : {
          "$ref" : "#/definitions/Adresse"
        },
        "commandeFournisseurList" : {
          "type" : "array",
          "items" : {
            "$ref" : "#/definitions/CommandeFournisseur"
          }
        },
        "entrepriseId" : {
          "type" : "integer",
          "format" : "int64"
        }
      }
    },
...
    "LigneCommandeClient" : {
      "type" : "object",
      "properties" : {
        "createdBy" : {
          "type" : "string"
        },
        "updatedBy" : {
          "type" : "string"
        },
        "createdDate" : {
          "type" : "string",
          "format" : "date-time"
        },
        "updateDate" : {
          "type" : "string",
          "format" : "date-time"
        },
        "id" : {
          "type" : "integer",
          "format" : "int64"
        },
        "article" : {
          "$ref" : "#/definitions/Article"
        },
        "commandeClient" : {
          "$ref" : "#/definitions/CommandeClient"
        },
        "entrepriseId" : {
          "type" : "integer",
          "format" : "int64"
        }
      }
    },
    "LigneCommandeClientDto" : {
      "type" : "object",
      "properties" : {
        "id" : {
          "type" : "integer",
          "format" : "int64"
        },
        "article" : {
          "$ref" : "#/definitions/ArticleDto"
        },
        "commandeClient" : {
          "$ref" : "#/definitions/CommandeClientDto"
        },
        "entrepriseId" : {
          "type" : "integer",
          "format" : "int64"
        }
      }
    },
    "LigneCommandeFournisseur" : {
      "type" : "object",
      "properties" : {
        "createdBy" : {
          "type" : "string"
        },
        "updatedBy" : {
          "type" : "string"
        },
        "createdDate" : {
          "type" : "string",
          "format" : "date-time"
        },
        "updateDate" : {
          "type" : "string",
          "format" : "date-time"
        },
        "id" : {
          "type" : "integer",
          "format" : "int64"
        },
        "article" : {
          "$ref" : "#/definitions/Article"
        },
        "commandeFournisseur" : {
          "$ref" : "#/definitions/CommandeFournisseur"
        },
        "entrepriseId" : {
          "type" : "integer",
          "format" : "int64"
        }
      }
    },
    "LigneCommandeFournisseurDto" : {
      "type" : "object",
      "properties" : {
        "id" : {
          "type" : "integer",
          "format" : "int64"
        },
        "article" : {
          "$ref" : "#/definitions/ArticleDto"
        },
        "commandeFournisseur" : {
          "$ref" : "#/definitions/CommandeFournisseurDto"
        },
        "entrepriseId" : {
          "type" : "integer",
          "format" : "int64"
        }
      }
    },
    "LigneVenteDto" : {
      "type" : "object",
      "properties" : {
        "id" : {
          "type" : "integer",
          "format" : "int64"
        },
        "article" : {
          "$ref" : "#/definitions/ArticleDto"
        },
        "vente" : {
          "$ref" : "#/definitions/VenteDto"
        },
        "quantite" : {
          "type" : "number"
        },
        "prixUniataire" : {
          "type" : "number"
        },
        "entrepriseId" : {
          "type" : "integer",
          "format" : "int64"
        }
      }
    },
    "RolesDto" : {
      "type" : "object",
      "properties" : {
        "id" : {
          "type" : "integer",
          "format" : "int64"
        },
        "name" : {
          "type" : "string",
          "enum" : [ "ROLE_USER", "ROLE_MODERATOR", "ROLE_ADMIN" ]
        }
      }
    },
    "SousCategorieDto" : {
      "type" : "object",
      "properties" : {
        "id" : {
          "type" : "integer",
          "format" : "int64"
        },
        "code" : {
          "type" : "string"
        },
        "libelle" : {
          "type" : "string"
        },
        "categorie" : {
          "$ref" : "#/definitions/CategorieDto"
        },
        "entrepriseId" : {
          "type" : "integer",
          "format" : "int64"
        }
      }
    },
    "UsersDto" : {
      "type" : "object",
      "properties" : {
        "id" : {
          "type" : "integer",
          "format" : "int64"
        },
        "username" : {
          "type" : "string"
        },
        "email" : {
          "type" : "string"
        },
        "password" : {
          "type" : "string"
        },
        "nom" : {
          "type" : "string"
        },
        "prenom" : {
          "type" : "string"
        },
        "dateNaissance" : {
          "type" : "integer",
          "format" : "int64"
        },
        "genre" : {
          "type" : "string"
        },
        "photo" : {
          "type" : "string"
        },
        "adresse" : {
          "$ref" : "#/definitions/AdresseDto"
        },
        "entreprise" : {
          "$ref" : "#/definitions/EntrepriseDto"
        },
        "roles" : {
          "type" : "array",
          "uniqueItems" : true,
          "items" : {
            "$ref" : "#/definitions/RolesDto"
          }
        }
      }
    },
    "VenteDto" : {
      "type" : "object",
      "properties" : {
        "id" : {
          "type" : "integer",
          "format" : "int64"
        },
        "code" : {
          "type" : "string"
        },
        "dateVente" : {
          "type" : "integer",
          "format" : "int64"
        },
        "commentaire" : {
          "type" : "string"
        },
        "entrepriseId" : {
          "type" : "integer",
          "format" : "int64"
        },
        "ligneVenteList" : {
          "type" : "array",
          "items" : {
            "$ref" : "#/definitions/LigneVenteDto"
          }
        }
      }
    }
  }
}